erf (erf v0.0.1)

erf is a library that provides a design-first framework to build RESTful APIs in Erlang.

Types

api/0

-type api() :: erf_parser:api().

body/0

-type body() :: njson:t().

conf/0

-type conf() ::
    #{spec_path := binary(),
      spec_parser => module(),
      callback := module(),
      preprocess_middlewares => [module()],
      postprocess_middlewares => [module()],
      port => inet:port_number(),
      ssl => boolean(),
      certfile => binary(),
      keyfile => binary(),
      static_routes => [static_route()],
      swagger_ui => boolean(),
      name => atom(),
      min_acceptors => pos_integer(),
      accept_timeout => pos_integer(),
      request_timeout => pos_integer(),
      header_timeout => pos_integer(),
      body_timeout => pos_integer(),
      max_body_size => pos_integer(),
      log_level => logger:level()}.

header/0

-type header() :: {binary(), binary()}.

method/0

-type method() :: get | post | put | delete | patch | head | options | trace | connect.

path_parameter/0

-type path_parameter() :: {binary(), binary()}.

query_parameter/0

-type query_parameter() :: {binary(), binary()}.

request/0

-type request() ::
    {Path :: [binary()],
     Method :: method(),
     QueryParameters :: [query_parameter()],
     Headers :: [header()],
     Body :: body()}.

response/0

-type response() ::
    {StatusCode :: pos_integer(), Headers :: [header()], Body :: body() | {file, binary()}}.

static_dir/0

-type static_dir() :: {dir, binary()}.

static_file/0

-type static_file() :: {file, binary()}.

static_route/0

-type static_route() :: {Path :: binary(), Resource :: static_file() | static_dir()}.

Functions

start_link(Conf)

-spec start_link(Conf) -> Result
              when
                  Conf :: conf(),
                  Result :: {ok, Pid} | ignore | {error, Reason},
                  Pid :: pid(),
                  Reason :: term().

Starts the supervision tree for an instance of the server.

stop(Name)

-spec stop(Name) -> ok when Name :: atom().

Stops the supervision tree for an instance of the server.