erf_router (erf v0.1.2)

Types

callback/0

-type callback() :: module().

generator_opts/0

-type generator_opts() :: #{callback := callback(), static_routes := [erf:static_route()]}.

t/0

-type t() :: erl_syntax:syntaxTree().

Functions

generate(API, Opts)

-spec generate(API, Opts) -> Result
                  when
                      API :: erf:api(),
                      Opts :: generator_opts(),
                      Result :: {Mod, Router},
                      Mod :: module(),
                      Router :: t().

Generates an Erlang Syntax Tree of a router module from an API AST.

handle(Name, Request)

-spec handle(Name, Request) -> Result
                when Name :: atom(), Request :: erf:request(), Result :: erf:response().

Handles an HTTP request.

load(Router)

-spec load(Router) -> Result
              when
                  Router :: t(),
                  Result :: ok | {ok, Warnings} | error | {error, {Errors, Warnings}},
                  Errors :: [term()],
                  Warnings :: [term()].

Loads a router module into the Erlang Runtime System.