erf_util (erf v0.1.2)

Functions

to_pascal_case(BinOrStr)

-spec to_pascal_case(BinOrStr) -> PascalCase
                  when BinOrStr :: binary() | string(), PascalCase :: binary().

Naive function to convert a binary or string to PascalCase.

to_snake_case(BinOrStr)

-spec to_snake_case(BinOrStr) -> SnakeCase when BinOrStr :: binary() | string(), SnakeCase :: binary().

Naive function to convert a binary or string to snake_case.